[MOB-16085] -- Finish mapping of geminio recipe
