+++ shared/features/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/GeminioSdkConstants.kt
@@ -3,11 +3,6 @@ package ru.hh.plugins.geminio.sdk
 
 internal object GeminioSdkConstants {
 
-    const val DEFAULT_REVISION_VALUE = 1
-    const val DEFAULT_MIN_API_VALUE = 1
-    const val DEFAULT_MIN_BUILD_API_VALUE = 1
-
-
     object HardcodedParams {
         /**
          * Package name from SELECTED file.
