--- shared/features/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/geminioRecipe.kt
@@ -1,18 +1,0 @@
-package ru.hh.plugins.geminio.sdk.template
-
-import com.android.tools.idea.wizard.template.RecipeExecutor
-import ru.hh.plugins.geminio.sdk.recipe.models.GeminioRecipe
-import ru.hh.plugins.geminio.sdk.template.models.GeminioRecipeExecutorData
-import ru.hh.plugins.geminio.sdk.template.executors.executeCommands
-
-
-fun RecipeExecutor.geminioRecipe(
-    geminioRecipe: GeminioRecipe,
-    executorData: GeminioRecipeExecutorData
-) {
-    executeCommands(geminioRecipe.recipeCommands, executorData)
-}
-
-
-
-
