+++ shared/features/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/template/mapping/GeminioRecipeMapper.kt
@@ -1,24 +1,54 @@
 package ru.hh.plugins.geminio.sdk.template.mapping
 
+import com.android.tools.idea.wizard.template.ModuleTemplateData
 import com.android.tools.idea.wizard.template.template
+import com.intellij.openapi.project.Project
+import ru.hh.plugins.freemarker_wrapper.FreemarkerConfiguration
+import ru.hh.plugins.geminio.sdk.GeminioSdkConstants
 import ru.hh.plugins.geminio.sdk.recipe.models.GeminioRecipe
 import ru.hh.plugins.geminio.sdk.template.aliases.AndroidStudioTemplate
+import ru.hh.plugins.geminio.sdk.template.executors.executeGeminioRecipe
 import ru.hh.plugins.geminio.sdk.template.mapping.optional.injectOptionalParams
 import ru.hh.plugins.geminio.sdk.template.mapping.required.injectRequiredParams
 import ru.hh.plugins.geminio.sdk.template.mapping.widgets.injectWidgets
+import ru.hh.plugins.geminio.sdk.template.models.GeminioRecipeExecutorData
 
 
 /**
  * Mapping from [ru.hh.plugins.geminio.sdk.recipe.models.GeminioRecipe]
  * into [ru.hh.plugins.geminio.sdk.template.aliases.AndroidStudioTemplate].
  */
-internal fun GeminioRecipe.toAndroidStudioTemplate(): AndroidStudioTemplate {
-    val recipe = this
+internal fun GeminioRecipe.toAndroidStudioTemplate(project: Project): AndroidStudioTemplate {
+    val geminioRecipe = this
 
     return template {
-        injectRequiredParams(recipe)
-        injectOptionalParams(recipe)
+        injectRequiredParams(geminioRecipe)
+        injectOptionalParams(geminioRecipe)
 
-        val existingParametersMap = injectWidgets(recipe)
+        val existingParametersMap = injectWidgets(geminioRecipe)
+
+        var isDryRun = true
+        recipe = { templateData ->
+            val moduleTemplateData = templateData as ModuleTemplateData
+            executeGeminioRecipe(
+                geminioRecipe = geminioRecipe,
+                executorData = GeminioRecipeExecutorData(
+                    project = project,
+                    isDryRun = isDryRun,
+                    moduleTemplateData = moduleTemplateData,
+                    existingParametersMap = existingParametersMap,
+                    resolvedParamsMap = existingParametersMap.asIterable().associate { entry ->
+                        entry.key to entry.value.value
+                    }.plus(
+                        mapOf(
+                            GeminioSdkConstants.HardcodedParams.PACKAGE_NAME to moduleTemplateData.packageName,
+                            GeminioSdkConstants.HardcodedParams.APPLICATION_PACKAGE to moduleTemplateData.projectTemplateData.applicationPackage
+                        )
+                    ),
+                    freemarkerConfiguration = FreemarkerConfiguration(geminioRecipe.freemarkerTemplatesRootDirPath)
+                )
+            )
+            isDryRun = false
+        }
     }
 }
